feat: :sparkles: Updated Gridify (Final)
