+++ changelog-md/v0-6-0.md
@@ -5,15 +5,10 @@ desc: 'Music previews and Gridify! 🎉'
 
 ### Added
 
--   **BETA** - Added Gridify! A new way to share your top songs in a grid format. You can find it by clicking on Gridify in the navigation bar!
+-   **BETA** - Added Gridify! A new way to share your top songs in a grid format. Export the image and share it on your social media! You can find it by clicking on Gridify in the menu found in the navigation bar!
+    -   Gridify is only available for Musicn users! If you want to use it, sign up for Musicn!
     -   You can click on the "Save" button in the page to save as an image and share it on your Social Media accounts.
 -   Added a new Music Preview. Click on any song on anybody's page to see a preview of the song!
-
-### Unresolved issues
-
--   The output image of Gridify pushes the text to the bottom. This is a known issue and will be fixed soon.
--   When your top song is set to private, the output image will show nothing. This is a known issue and will be fixed soon.
-    -   The workaround is to temporarily set your top song to public and then set it back to private.
 
 ### Changed
